Tests/LibWeb/Text/expected/hit_testing/nested-inline-paintables.txt

   lang "en"    <SPAN id="name" >
<SPAN id="value" >
